+++ client/src/routes/+layout@base.svelte
