+++ client/src/routes/about/+page.svelte
@@ -1,17 +1,19 @@
 <script context="module">
-  import { browser, dev } from "$app/env";
+  throw new Error("@migration task: Check code was safely removed (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292722)");
 
-  // we don't need any JS on this page, though we'll load
-  // it in dev so that we get hot module replacement...
-  export const hydrate = dev;
+  // import { browser, dev } from "$app/env";
 
-  // ...but if the client-side router is already loaded
-  // (i.e. we came here from elsewhere in the app), use it
-  export const router = browser;
+  // // we don't need any JS on this page, though we'll load
+  // // it in dev so that we get hot module replacement...
+  // export const hydrate = dev;
 
-  // since there's no dynamic data here, we can prerender
-  // it so that it gets served as a static asset in prod
-  export const prerender = true;
+  // // ...but if the client-side router is already loaded
+  // // (i.e. we came here from elsewhere in the app), use it
+  // export const router = browser;
+
+  // // since there's no dynamic data here, we can prerender
+  // // it so that it gets served as a static asset in prod
+  // export const prerender = true;
 </script>
 
 <svelte:head>
