+++ client/src/routes/fiches/[id].spec/+server.ts
@@ -1,3 +1,5 @@
+
+// @migration task: Check imports
 /**
  * @jest-environment jsdom
  */
