+++ client/src/routes/fiches/search/+page.svelte
@@ -1,44 +1,48 @@
 <script context="module" lang="ts">
-  import type { Load } from "@sveltejs/kit";
-  import { get } from "svelte/store";
-  import { getDatasets } from "$lib/repositories/datasets";
-  import { apiToken } from "$lib/stores/auth";
-  import { getPageFromParams } from "$lib/util/pagination";
-  import { page as pageStore } from "$app/stores";
-  import { getDatasetFiltersInfo } from "src/lib/repositories/datasetFilters";
-  import { toFiltersValue } from "src/lib/transformers/datasetFilters";
+  throw new Error("@migration task: Check code was safely removed (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292722)");
 
-  export const load: Load = async ({ fetch, url }) => {
-    const page = getPageFromParams(url.searchParams);
-    const q = url.searchParams.get("q") || "";
-    const filtersValue = toFiltersValue(url.searchParams);
+  // import type { Load } from "@sveltejs/kit";
+  // import { get } from "svelte/store";
+  // import { getDatasets } from "$lib/repositories/datasets";
+  // import { apiToken } from "$lib/stores/auth";
+  // import { getPageFromParams } from "$lib/util/pagination";
+  // import { page as pageStore } from "$app/stores";
+  // import { getDatasetFiltersInfo } from "src/lib/repositories/datasetFilters";
+  // import { toFiltersValue } from "src/lib/transformers/datasetFilters";
 
-    const token = get(apiToken);
+  // export const load: Load = async ({ fetch, url }) => {
+  //   const page = getPageFromParams(url.searchParams);
+  //   const q = url.searchParams.get("q") || "";
+  //   const filtersValue = toFiltersValue(url.searchParams);
 
-    const [paginatedDatasets, filtersInfo] = await Promise.all([
-      getDatasets({
-        fetch,
-        apiToken: token,
-        page,
-        q,
-        filters: filtersValue,
-      }),
-      getDatasetFiltersInfo({ fetch, apiToken: token }),
-    ]);
+  //   const token = get(apiToken);
 
-    return {
-      props: {
-        paginatedDatasets,
-        filtersInfo,
-        filtersValue,
-        currentPage: page,
-        q,
-      },
-    };
-  };
+  //   const [paginatedDatasets, filtersInfo] = await Promise.all([
+  //     getDatasets({
+  //       fetch,
+  //       apiToken: token,
+  //       page,
+  //       q,
+  //       filters: filtersValue,
+  //     }),
+  //     getDatasetFiltersInfo({ fetch, apiToken: token }),
+  //   ]);
+
+  //   return {
+  //     props: {
+  //       paginatedDatasets,
+  //       filtersInfo,
+  //       filtersValue,
+  //       currentPage: page,
+  //       q,
+  //     },
+  //   };
+  // };
 </script>
 
 <script lang="ts">
+  throw new Error("@migration task: Add data prop (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292707)");
+
   import { goto } from "$app/navigation";
   import type { Dataset } from "src/definitions/datasets";
   import type {
@@ -51,10 +55,10 @@
   import { Maybe } from "$lib/util/maybe";
   import { pluralize } from "src/lib/util/format";
   import type { Paginated } from "src/definitions/pagination";
-  import FilterPanel from "./_FilterPanel.svelte";
+  import FilterPanel from "../_FilterPanel.svelte";
   import { toFiltersParams } from "src/lib/transformers/datasetFilters";
   import { makePageParam } from "$lib/util/pagination";
-  import PaginationContainer from "./_PaginationContainer.svelte";
+  import PaginationContainer from "../_PaginationContainer.svelte";
 
   export let paginatedDatasets: Maybe<Paginated<Dataset>>;
   export let q: string;
