+++ client/src/routes/login/+page.svelte
@@ -1,7 +1,3 @@
-<script lang="ts" context="module">
-  export const prerender = true;
-</script>
-
 <script lang="ts">
   import { goto } from "$app/navigation";
   import type { LoginFormData } from "src/definitions/auth";
